completed functionality of the history/favorites component
