+++ client/src/dashboard/TrendrArticle.js
@@ -40,12 +40,7 @@ const useStyles = makeStyles((theme) => ({
 export default function TrendrArticle(props) {
   const classes = useStyles();
 
-  const saveHistory = (headline, url) => {
-    axios.post("/save", {headline, url})
-    .then((res) => {
-      console.log(res)
-    })
-  }
+  
 
   return (
 
@@ -55,7 +50,7 @@ export default function TrendrArticle(props) {
       {props.data ? (
 
         <Table key={props.data.article.articleTitle}>
-          <TableBody className={classes.table}><img className={classes.borderedImg} src={props.data.image}></img><StarBorderIcon className={classes.favoritesIcon} onClick={() => {saveHistory(props.data.article.articleTitle, props.data.article.url)}}/></TableBody>
+          <TableBody className={classes.table}><img className={classes.borderedImg} src={props.data.image}></img><StarBorderIcon className={classes.favoritesIcon} onClick={() => {props.saveHistory(props.data.article.articleTitle, props.data.article.url)}}/></TableBody>
           <br></br>
           <TableBody>{props.data.article.articleTitle.replace(/&#39;/g, "'").replace("&amp;", "&").replace(/&quot;/g, '"')}</TableBody>
           <TableBody>{props.data.article.snippet.replace(/&#39;/g, "'").replace("&amp;", "&").replace(/&quot;/g, '"')}</TableBody>
